requirements.txt
yt-dlp
opencv-python
numpy
pygame
Pillow
PyYAML
